# sha: use EVP_DIGEST instead of go_shaX