This PR is a clean-up opportunity I saw recently. We don't need to implement a one-shot digest function ourselves, all supported OpenSSL versions already provide [EVP_Digest](https://www.openssl.org/docs/man3.0/man3/EVP_Digest.html). This function has the nice benefit of being a little bit faster on OpenSSL 3, thanks to setting the `EVP_MD_CTX_FLAG_ONESHOT` flag.

```cmd
goos: linux
goarch: amd64
pkg: github.com/golang-fips/openssl
cpu: AMD EPYC 7763 64-Core Processor                
          │   new.txt   │              new2.txt              │
          │   sec/op    │   sec/op     vs base               │
SHA256-16   287.9n ± 1%   283.8n ± 1%  -1.46% (p=0.003 n=10)

          │   new.txt    │              new2.txt               │
          │     B/s      │     B/s       vs base               │
SHA256-16   26.49Mi ± 1%   26.89Mi ± 1%  +1.49% (p=0.002 n=10)
¹ all samples are equal
```